moved stream around
