+++ CMakeFiles/Makefile.cmake
@@ -10,95 +10,25 @@ set(CMAKE_MAKEFILE_DEPENDS
   "CMakeFiles/3.9.1/CMakeCCompiler.cmake"
   "CMakeFiles/3.9.1/CMakeCXXCompiler.cmake"
   "CMakeFiles/3.9.1/CMakeSystem.cmake"
-  "CMakeFiles/feature_tests.c"
-  "CMakeFiles/feature_tests.cxx"
   "CMakeLists.txt"
   "libsgp4/CMakeLists.txt"
-  "/usr/share/cmake-3.9/Modules/CMakeCCompiler.cmake.in"
-  "/usr/share/cmake-3.9/Modules/CMakeCCompilerABI.c"
   "/usr/share/cmake-3.9/Modules/CMakeCInformation.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeCXXCompiler.cmake.in"
-  "/usr/share/cmake-3.9/Modules/CMakeCXXCompilerABI.cpp"
   "/usr/share/cmake-3.9/Modules/CMakeCXXInformation.cmake"
   "/usr/share/cmake-3.9/Modules/CMakeCommonLanguageInclude.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeCompilerIdDetection.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeConfigurableFile.in"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineCCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineCXXCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineCompileFeatures.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineCompilerABI.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineCompilerId.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeDetermineSystem.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeFindBinUtils.cmake"
   "/usr/share/cmake-3.9/Modules/CMakeGenericSystem.cmake"
   "/usr/share/cmake-3.9/Modules/CMakeLanguageInformation.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeParseImplicitLinkInfo.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeSystem.cmake.in"
   "/usr/share/cmake-3.9/Modules/CMakeSystemSpecificInformation.cmake"
   "/usr/share/cmake-3.9/Modules/CMakeSystemSpecificInitialize.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeTestCCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeTestCXXCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeTestCompilerCommon.cmake"
-  "/usr/share/cmake-3.9/Modules/CMakeUnixFindMake.cmake"
-  "/usr/share/cmake-3.9/Modules/CheckFunctionExists.c"
-  "/usr/share/cmake-3.9/Modules/CheckIncludeFile.c.in"
   "/usr/share/cmake-3.9/Modules/CheckIncludeFile.cmake"
   "/usr/share/cmake-3.9/Modules/CheckLibraryExists.cmake"
   "/usr/share/cmake-3.9/Modules/CheckSymbolExists.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/ADSP-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/ARMCC-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/AppleClang-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Borland-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Bruce-C-DetermineCompiler.cmake"
   "/usr/share/cmake-3.9/Modules/Compiler/CMakeCommonCompilerMacros.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Clang-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Clang-DetermineCompilerInternal.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Comeau-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Compaq-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Compaq-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Cray-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Embarcadero-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Fujitsu-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/GHS-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/GNU-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/GNU-C-FeatureTests.cmake"
   "/usr/share/cmake-3.9/Modules/Compiler/GNU-C.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/GNU-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/GNU-CXX-FeatureTests.cmake"
   "/usr/share/cmake-3.9/Modules/Compiler/GNU-CXX.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/GNU-FindBinUtils.cmake"
   "/usr/share/cmake-3.9/Modules/Compiler/GNU.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/HP-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/HP-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/IAR-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/IBMCPP-C-DetermineVersionInternal.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/IBMCPP-CXX-DetermineVersionInternal.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Intel-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/MIPSpro-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/MSVC-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/NVIDIA-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/OpenWatcom-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/PGI-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/PathScale-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/SCO-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/SDCC-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/SunPro-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/SunPro-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/TI-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/TinyCC-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/VisualAge-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/VisualAge-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/Watcom-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/XL-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/XL-CXX-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/zOS-C-DetermineCompiler.cmake"
-  "/usr/share/cmake-3.9/Modules/Compiler/zOS-CXX-DetermineCompiler.cmake"
   "/usr/share/cmake-3.9/Modules/FindPackageHandleStandardArgs.cmake"
   "/usr/share/cmake-3.9/Modules/FindPackageMessage.cmake"
   "/usr/share/cmake-3.9/Modules/FindThreads.cmake"
-  "/usr/share/cmake-3.9/Modules/Internal/FeatureTesting.cmake"
-  "/usr/share/cmake-3.9/Modules/Platform/Linux-Determine-CXX.cmake"
   "/usr/share/cmake-3.9/Modules/Platform/Linux-GNU-C.cmake"
   "/usr/share/cmake-3.9/Modules/Platform/Linux-GNU-CXX.cmake"
   "/usr/share/cmake-3.9/Modules/Platform/Linux-GNU.cmake"
@@ -114,11 +44,6 @@ set(CMAKE_MAKEFILE_OUTPUTS
 
 # Byproducts of CMake generate step:
 set(CMAKE_MAKEFILE_PRODUCTS
-  "CMakeFiles/3.9.1/CMakeSystem.cmake"
-  "CMakeFiles/3.9.1/CMakeCCompiler.cmake"
-  "CMakeFiles/3.9.1/CMakeCXXCompiler.cmake"
-  "CMakeFiles/3.9.1/CMakeCCompiler.cmake"
-  "CMakeFiles/3.9.1/CMakeCXXCompiler.cmake"
   "CMakeFiles/CMakeDirectoryInformation.cmake"
   "libsgp4/CMakeFiles/CMakeDirectoryInformation.cmake"
   )
